Implemented custom hook to manage task state data
